Added download section to resume page
Added buttons to download the resume as a .docx, a .pdf, and to simply print the file. At the moment, the save buttons and the print button all save/print the same file: test.pdf. Given the react-pdf libraries nature of rendering the document, everything on this page has been manually placed to avoid breaking the document rendering. The page still needs to be refactored to accomadate smaller screens.