Theme change button, first project edited
Added a theme change button to the bottom of the navbar. Currently missing functionality, animations, and it needs to be resized so that it does not close when mouse moves off the button. Added description and some other bullet points to the first project in projects, and made a second copy of the template